Modified types from unsigned int to size_t in bit/array
git-svn-id: http://svn.ta3d.org/yuni/trunk@1016 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/yuni/core/bit/array.hxx
@@ -28,20 +28,20 @@ namespace Bit
 
 	inline void Array::unset()
 	{
-		(void)::memset(pBuffer.data(), 0, (unsigned int)pBuffer.sizeInBytes());
+		(void)::memset(pBuffer.data(), 0, (size_t)pBuffer.sizeInBytes());
 	}
 
 
 
 	inline void Array::reset()
 	{
-		(void)::memset(pBuffer.data(), 0, (unsigned int)pBuffer.sizeInBytes());
+		(void)::memset(pBuffer.data(), 0, (size_t)pBuffer.sizeInBytes());
 	}
 
 
 	inline void Array::reset(bool value)
 	{
-		(void)::memset(pBuffer.data(), (value ? 0xFF : 0), (unsigned int)pBuffer.sizeInBytes());
+		(void)::memset(pBuffer.data(), (value ? 0xFF : 0), (size_t)pBuffer.sizeInBytes());
 	}
